+++ seed/ts-sdk/literal/src/api/resources/reference/client/Client.ts
@@ -50,7 +50,8 @@ export class Reference {
             method: "POST",
             headers: {
                 "X-API-Version": requestOptions?.version ?? "02-02-2024",
-                "X-API-Enable-Audit-Logging": requestOptions?.auditLogging ?? "true",
+                "X-API-Enable-Audit-Logging":
+                    requestOptions?.auditLogging != null ? requestOptions.auditLogging.toString() : "true",
                 "X-Fern-Language": "JavaScript",
                 "X-Fern-SDK-Name": "@fern/literal",
                 "X-Fern-SDK-Version": "0.0.1",